1.优化代码 2.添加分类标签删除时有文章不能删除的逻辑

+++ blog/common/src/main/java/com/shiyi/service/impl/CategoryServiceImpl.java
@@ -5,8 +5,10 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.shiyi.common.FieldConstants;
 import com.shiyi.common.ResponseResult;
+import com.shiyi.entity.Article;
 import com.shiyi.entity.Category;
 import com.shiyi.exception.BusinessException;
+import com.shiyi.mapper.ArticleMapper;
 import com.shiyi.mapper.CategoryMapper;
 import com.shiyi.service.CategoryService;
 import com.shiyi.utils.DateUtil;
@@ -17,6 +19,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -35,6 +38,8 @@ import static com.shiyi.common.ResultCode.CATEGORY_IS_TOP;
 @Service
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> implements CategoryService {
+
+    private final ArticleMapper articleMapper;
 
     /**
      * 分类列表
@@ -98,6 +103,10 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     @Override
     @Transactional(rollbackFor = Exception.class)
     public ResponseResult deleteCategory(Long id) {
+        Integer count = articleMapper.selectCount(new QueryWrapper<Article>().eq(FieldConstants.CATEGORY_ID, id));
+        if (count > 0) {
+            throw new BusinessException( "分类下有文章，不能删除");
+        }
         int rows = baseMapper.deleteById(id);
         return rows > 0 ? ResponseResult.success(): ResponseResult.error("删除分类失败");
     }
@@ -110,7 +119,15 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     @Override
     @Transactional(rollbackFor = Exception.class)
     public ResponseResult deleteBatch(List<Category> list) {
-        List<Long> ids = list.stream().map(Category::getId).collect(Collectors.toList());
+        List<Long> ids = new ArrayList<>();
+        for (Category category : list) {
+            Integer count = articleMapper.selectCount(new QueryWrapper<Article>().eq(FieldConstants.CATEGORY_ID, category.getId()));
+            if (count > 0) {
+                throw new BusinessException( category.getName() + " 分类下有文章，不能删除");
+            }
+            ids.add(category.getId());
+        }
+
         int rows = baseMapper.deleteBatchIds(ids);
         return rows > 0 ? ResponseResult.success(): ResponseResult.error("批量删除分类失败");
     }
